pyproject.toml
[build-system]
requires = ["hatchling>=1.21,<1.22"]
build-backend = "hatchling.build"

[project]
name = "agentanycast"
version = "0.3.0"
description = "Python SDK for AgentAnycast — A2A protocol over P2P"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    { name = "AgentAnycast Contributors" },
]
keywords = [
    "a2a", "a2a-protocol", "agent", "agent-communication", "p2p", "libp2p",
    "grpc", "mcp", "crewai", "langgraph", "google-adk", "openai-agents", "zero-config",
    "end-to-end-encryption", "nat-traversal", "decentralized", "noise-protocol", "did",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "grpcio>=1.60",
    "protobuf>=4.25",
    "httpx>=0.27",
    "click>=8.0",
    "base58>=2.1",
]

[project.urls]
Homepage = "https://github.com/AgentAnycast/agentanycast"
Repository = "https://github.com/AgentAnycast/agentanycast-python"
Issues = "https://github.com/AgentAnycast/agentanycast/issues"
Changelog = "https://github.com/AgentAnycast/agentanycast-python/blob/main/CHANGELOG.md"

[project.scripts]
agentanycast = "agentanycast.cli.main:main"

[project.optional-dependencies]
crewai = ["crewai>=0.80"]
langgraph = ["langgraph>=0.2"]
adk = ["google-adk>=1.0"]
openai-agents = ["openai-agents>=0.10"]
adapters = ["crewai>=0.80", "langgraph>=0.2", "google-adk>=1.0", "openai-agents>=0.10"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agentanycast"]

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = ["src/agentanycast/_generated"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = "agentanycast._generated.*"
ignore_errors = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = "base58"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["grpc", "grpc.*", "grpc.aio", "grpc.aio.*", "google.protobuf.*", "crewai", "crewai.*", "langgraph", "langgraph.*", "google.adk", "google.adk.*", "google.genai", "google.genai.*", "agents", "agents.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: mark test as integration test (requires relay/daemon binaries)",
]
# Default: skip integration tests. Run explicitly with: pytest -m integration
addopts = "-m 'not integration'"
